Refactor: Move project close logic to EditorHandlerActivity
The logic for handling project closure, including confirming with the user about unsaved changes and performing the close operation, has been moved from `ProjectHandlerActivity` to `EditorHandlerActivity`.

Specifically:
- The `confirmProjectClose()` method and its related dialog handling are now in `EditorHandlerActivity`.
- A new private method `performCloseAllFiles(manualFinish: Boolean)` is introduced in `EditorHandlerActivity` to encapsulate the core logic of closing all files and plugin tabs. This method is called by both the refactored `closeAll()` and the new `confirmProjectClose()`.
- The `doCloseAll()` abstract method in `ProjectHandlerActivity` no longer takes a `runAfter` lambda.
- Calls to `closeProject(true)` in `ProjectHandlerActivity` are replaced with `doCloseAll()` or `doConfirmProjectClose()` as appropriate.
- The `closeAll(runAfter: () -> Unit)` in `EditorHandlerActivity` now delegates to `performCloseAllFiles(manualFinish = false)` after handling unsaved files, and then executes `runAfter`.
- When closing a project, `GeneralPreferences.lastOpenedProject` is now set to `GeneralPreferences.NO_OPENED_PROJECT` within `EditorHandlerActivity`'s `confirmProjectClose` logic, after the user confirms and before files are actually closed.
