+++ app/src/main/java/com/itsaky/androidide/activities/editor/ProjectHandlerActivity.kt
@@ -164,7 +164,7 @@ abstract class ProjectHandlerActivity : BaseEditorActivity() {
         const val STATE_KEY_SHOULD_INITIALIZE = "ide.editor.isInitializing"
     }
 
-    abstract fun doCloseAll(runAfter: () -> Unit)
+    abstract fun doCloseAll()
 
     abstract fun saveOpenedFiles()
 
@@ -172,10 +172,6 @@ abstract class ProjectHandlerActivity : BaseEditorActivity() {
         if (mSearchingProgress?.isShowing == true) {
             mSearchingProgress!!.dismiss()
         }
-    }
-
-    override fun doConfirmProjectClose() {
-        confirmProjectClose()
     }
 
     override fun doOpenHelp() {
@@ -330,7 +326,7 @@ abstract class ProjectHandlerActivity : BaseEditorActivity() {
             this.initializingFuture?.cancel(true)
             this.initializingFuture = null
 
-            closeProject(false)
+            doCloseAll()
         }
 
         if (IDELanguageClientImpl.isInitialized()) {
@@ -877,60 +873,10 @@ abstract class ProjectHandlerActivity : BaseEditorActivity() {
         }
     }
 
-    private fun closeProject(manualFinish: Boolean) {
-        if (manualFinish) {
-            // if the user is manually closing the project,
-            // save the opened files cache
-            // this is needed because in this case, the opened files cache will be empty
-            // when onPause will be called.
-            saveOpenedFiles()
-
-            // reset the lastOpenedProject if the user explicitly chose to close the project
-            GeneralPreferences.lastOpenedProject = GeneralPreferences.NO_OPENED_PROJECT
-        }
-
-        // Make sure we close files
-        // This will make sure that file contents are not erased.
-        doCloseAll {
-            if (manualFinish) {
-                finish()
-            }
-        }
-    }
-
     private fun openHelpActivity() {
         val intent = Intent(this, HelpActivity::class.java)
         intent.putExtra(CONTENT_KEY, HELP_PAGE_URL)
         startActivity(intent)
-    }
-
-    private fun confirmProjectClose() {
-        val builder = newMaterialDialogBuilder(this)
-        builder.setTitle(string.title_confirm_project_close)
-        builder.setMessage(string.msg_confirm_project_close)
-
-        builder.setNegativeButton(string.cancel_project_text, null)
-
-        builder.setNeutralButton(string.close_without_saving) { dialog, _ ->
-            dialog.dismiss()
-
-            editorActivityScope.launch {
-                for (i in 0 until editorViewModel.getOpenedFileCount()) {
-                    (content.editorContainer.getChildAt(i) as? CodeEditorView)?.editor?.markUnmodified()
-                }
-
-                withContext(Dispatchers.Main) {
-                    closeProject(true)
-                }
-            }
-        }
-
-        builder.setPositiveButton(string.save_close_project) { dialog, _ ->
-            dialog.dismiss()
-            closeProject(true)
-        }
-
-        builder.show()
     }
 
     private fun initLspClient() {
